Record of thinking 3.
This past cycle was one of my post productive cycles. During this time
I was able to finish an improved version of the class. (I had to redo
my old class because it was not working properly) I also was able to
nearly finish the computer portion. What I noticed was that I needed a
lot more methods than I thought I would. EX: checking if it is a valid
space or not etc. Also upon rewriting the class, I decided to use a 5x5
array instead of a 3x3. This ended up being a pretty big improvement
and I am glad I did it because it helped me understand for loops
better. Overall, I was able to gain more knowledge about for loops, two
diminutional arrays and calling methods as well as strings.